fix: WorldAnchorの移動タイミングとTrackedImagesChangedObservableを使っている他の処理のタイミングが違う
